Fix replicate bits types, still fails the binary_operation left type == right type check
This is due to the `or` operation which is between two `Bits` types, but when emitting, these will be concrete sizes A and B, where A is the output width, and B is the width of the replicated section. Can't think of a good way of fixing this, maybe some runtime cast? But it works fine as is, will try and find another way to fix the `stp` encoder error